requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==3.10.1
Flask-Caching==2.3.0
SQLAlchemy==2.0.46
alembic==1.14.1
python-dotenv==1.0.1
email-validator==2.2.0
Pillow==11.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
sentry-sdk[flask]==2.19.2
reportlab==4.2.5
